# openapi.yaml
openapi: 3.1.0
info:
  title: LearnFast Core Engine
  description: Hybrid Graph-RAG Learning Platform API
  version: 0.1.0
paths:
...
  /api/concepts/unlocked/{user_id}:
    get:
      tags:
      - navigation
      summary: Get unlocked concepts for user
      description: Get concepts available for the user to start.
      operationId: get_unlocked_concepts_api_concepts_unlocked__user_id__get
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          title: User Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ConceptNode'
                title: Response Get Unlocked Concepts Api Concepts Unlocked  User
                  Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/learning/lesson/{user_id}/{target_concept}:
    get:
      tags:
      - navigation
      summary: Get formatted lesson content
      description: 'Get full lesson content (text) for a path to target.

        Convenience endpoint that resolves path + retrieves content.'
      operationId: get_lesson_content_api_learning_lesson__user_id___target_concept__get
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          title: User Id
      - name: target_concept
        in: path
        required: true
        schema:
          type: string
          title: Target Concept
      - name: time_budget
        in: query
        required: false
        schema:
          type: integer
          default: 30
          title: Time Budget
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/progress/start:
    post:
      tags:
      - navigation
      summary: Mark concept in progress
      description: Mark a concept as started (IN_PROGRESS).
      operationId: start_concept_api_progress_start_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProgressUpdate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/progress/complete:
    post:
      tags:
      - navigation
      summary: Mark concept completed
      description: Mark a concept as COMPLETED.
      operationId: complete_concept_api_progress_complete_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProgressUpdate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/progress/{user_id}:
    get:
      tags:
      - navigation
      summary: Get full user state
      description: Get user's completed, in-progress, and available concepts.
      operationId: get_user_progress_api_progress__user_id__get
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          title: User Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserState'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ConceptNode:
      properties:
        name:
          type: string
          title: Name
          description: Unique concept name (lowercase)
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
          description: Concept description
        depth_level:
          anyOf:
          - type: integer
          - type: 'null'
          title: Depth Level
          description: Depth in prerequisite hierarchy
      type: object
      required:
      - name
      title: ConceptNode
      description: Represents a concept node in the knowledge graph.
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ProgressUpdate:
      properties:
        user_id:
          type: string
          title: User Id
        concept_name:
          type: string
          title: Concept Name
      type: object
      required:
      - user_id
      - concept_name
      title: ProgressUpdate
      description: Request for updating concept progress.
...
    UserState:
      properties:
        user_id:
          type: string
          title: User Id
        completed_concepts:
          items:
            type: string
          type: array
          title: Completed Concepts
        in_progress_concepts:
          items:
            type: string
          type: array
          title: In Progress Concepts
        available_concepts:
          items:
            type: string
          type: array
          title: Available Concepts
      type: object
      required:
      - user_id
      - completed_concepts
      - in_progress_concepts
      - available_concepts
      title: UserState
      description: Represents current user progress and available options.
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError